added orientation check, docstrings, improved logging

+++ grabber/grab.py
@@ -2,6 +2,7 @@ from cv2 import IMREAD_COLOR
 from cv2 import imdecode
 from cv2 import imwrite
 from PIL import Image
+import improc.features.preprocess as preprocess
 
 import numpy as np
 
@@ -9,29 +10,61 @@ from os import mkdir
 from os import path
 from pymongo import MongoClient
 import requests
-import improc.features.preprocess as preprocess
 from StringIO import StringIO
 from collections import defaultdict
+import logging
+
 LOGGER = logging.getLogger(__name__)
-LOGGER.setLevel(logging.DEBUG)
+LOGGER.setLevel(logging.ERROR)
+logging.info('Starting logger for image grabber.')
+
+OUT_DIR = 'out'
+MAX_NUM_PRODUCTS = 400000
+
+EXCLUDED_IMAGE_ORIENTATIONS = [(270, 90),
+                               (270, 270),
+                               (180, 180),
+                               (90, 180)]
 
 def is_corrupted(stream):
+    """Check if the data streamed by url request is a valid image
+    :param stream: raw data from url
+    :return bool: whether it is a valid image or not
+    """
     try:
         img = Image.open(StringIO(stream.content))
         check = img.verify()
         del img
         return False
-    except IOError:
+    except IOError as e:
+        LOGGER.error(e)
+        return True
+
+
+def is_valid_orientation(img_dict):
+    """Check that the shoe is not in a forbidden orientation.
+    :param img_dict: image dictionary to test
+    :return: false if orientation (pitch, yaw) is in EXCLUDED_IMAGE_ORIENTATIONS
+    """
+    if ('y' in img_dict and 'z' in img_dict and
+            (img_dict['y'], img_dict['z']) in EXCLUDED_IMAGE_ORIENTATIONS):
+        return False
+    else:
         return True
 
 
 def image_raw_preprocessing(img_stream):
+    """Decode the raw data from url into an image, crops and makes square
+    :param img_stream: raw data from url
+    :return: processed img
+    """
     image_squared = None
     image_decoded = imdecode(np.fromstring(img_stream.content, np.uint8), flags=IMREAD_COLOR)
     if image_decoded is not None:
         try:
             image_autocropped = preprocess.autocrop(image_decoded)
-        except AttributeError:
+        except AttributeError as e:
+            LOGGER.error(e)
             return image_squared
         if image_autocropped is not None:
             image_scaled_max = preprocess.scale_max(image_autocropped)
@@ -40,6 +73,11 @@ def image_raw_preprocessing(img_stream):
 
 
 def update_product(data, prd_id):
+    """Updates a product in the collection with data stored in a dictionary
+    :param data: dictionary of new data
+    :param prd_id: id of the product to be updated
+    :return:
+    """
     if any(data):
         collection.update(
             {"_id": prd_id},
@@ -48,37 +86,23 @@ def update_product(data, prd_id):
         )
 
 
-out_dir = 'out'
-
-# Initializing MongoDB client
-client = MongoClient('localhost', 27017)
-test_db = client.jemboo
-collection = test_db.shoes
-
-if not (path.exists(out_dir)):
-    mkdir(out_dir)
-    print "Created output folder"
-
-products = collection.find().batch_size(50)
-
-print "Downloading images..."
-print_interval = products.count() / 20
-num_processed_products = 0
-counters = defaultdict(int)
-
-
-def process_image(set_data):
+def process_image(img, data):
+    """Checks if an image in the db has not been processed and has a valid
+       orientation, then processes the image.
+    :param img: A dictionary from the database which stores data about the
+                image to be processed.
+    :param data: A dictionary where to store image status for future update.
+    :return: The image status.
+    """
     try:
-        global product_status
         url = img['url']
         img_id = img['_id']
-        img_filename = out_dir + "/" + str(product_id) + "_" + str(img_id) + ".jpg"
-        if "image_processed_status" not in img:  # check if image already exist
-
-            img_data = requests.get(url, stream=True)
-            if img_data.status_code == 200:
-                if not is_corrupted(img_data):
-                    processed_img = image_raw_preprocessing(img_data)
+        img_filename = OUT_DIR + "/" + str(product_id) + "_" + str(img_id) + ".jpg"
+        if "image_processed_status" not in img and is_valid_orientation(img):  # check if image already exist
+            img_raw_data = requests.get(url, stream=True)
+            if img_raw_data.status_code == 200:
+                if not is_corrupted(img_raw_data):
+                    processed_img = image_raw_preprocessing(img_raw_data)
                     if processed_img is not None:
                         imwrite(img_filename, processed_img)  # save image
                         image_status = "ok"
@@ -92,7 +116,7 @@ def process_image(set_data):
                 print("Unable to retrieve image " + str(img_index) + "/" + str(prd_index))
                 image_status = "http_fail"
 
-            set_data["images.%s.image_processed_status" % img_index] = image_status  # update img status
+            data["images.%s.image_processed_status" % img_index] = image_status  # update img status
         else:
             image_status = "already_processed"
 
@@ -102,28 +126,46 @@ def process_image(set_data):
         return "exception"
 
 
-while counters["images_attempted"] < 400000:
-    for prd_index, product in enumerate(products):
+# Initializing MongoDB client
+client = MongoClient('localhost', 27017)
+test_db = client.jemboo
+collection = test_db.shoes
+
+if not (path.exists(OUT_DIR)):
+    mkdir(OUT_DIR)
+    print "Created output folder"
+
+products = collection.find().batch_size(50)
+
+print "Downloading images..."
+print_interval = min((products.count(), MAX_NUM_PRODUCTS)) / 50
+counters = defaultdict(int)
+
+for prd_index, product in enumerate(products):
+    if counters["images_attempted"] < MAX_NUM_PRODUCTS:
         product_status = "ok"
         product_id = product['_id']
 
         if prd_index % print_interval == 0:
-            print(str(prd_index / print_interval * 4) + "% of products scanned")
+            print("\r" + str(prd_index / print_interval * 2) + "% of products scanned")
 
-        set_data = {}
-        for img_index, img in enumerate(product['images']):
+        db_changes = {}
+        for img_index, img_data in enumerate(product['images']):
             counters["images_attempted"] += 1
-            processed_image_status = process_image(set_data)
+            processed_image_status = process_image(img_data, db_changes)
             counters[processed_image_status] += 1
+            if processed_image_status is not "ok":
+                product_status = "failed"
 
-        set_data["processed_status"] = product_status  # update_product_status
-        update_product(set_data, product_id)
-        num_processed_products += 1
+        db_changes["processed_status"] = product_status  # update_product_status
+        update_product(db_changes, product_id)
+        counters["products_attempted"] += 1
+    else:
+        break
 
-LOGGER.info("100% of products scanned\n")
-LOGGER.info("Total number of processed products: %i", num_processed_products)
+LOGGER.info("100% of products scanned.")
+LOGGER.info("Total number of processed products: %i", counters["products_attempted"])
 LOGGER.info("Total number of processed images: %i", counters["images_attempted"])
 LOGGER.info("Number of images failed to crop: %i", counters["autocropped_failed"])
-
 LOGGER.info("Number of corrupted images: %i", counters["image_corrupted"])
